skills/video-understand/pyproject.toml
[project]
name = "skill-video-understand"
version = "0.1.0"
description = "Understand video content with OpenAI-compatible multimodal models"
readme = "SKILL.md"
requires-python = ">=3.10"
dependencies = [
	"openai>=1.0.0",
	"python-dotenv>=1.0.0",
]

[tool.uv]
package = false
